requirements.txt
# Django-specific packages
Django==1.7

# Django 6 ok
django_compressor==1.4
django-licensing==0.2.4
djangorestframework==2.3.13
django-model-utils==2.0.3
django-reversion==1.8.4

# ???
django-browserid==0.7.1

pyelasticsearch==0.6
BeautifulSoup<4.0
lxml==3.0.2
cssselect==0.8
Pillow==2.3.0
psycopg2==2.4.6

isodate==0.4.9
pytz==2012h
refine-client==0.2.1
smartypants==1.6.0.3

ipython==2.3.0

webtest==2.0.14
django-webtest==1.7.6
